+++ src/__test__/define-component.test.js
