+++ src/define-component.js
