+++ src/pack.js
